Deduplicate the WASM interface structs
