+++ zellij-tile/Cargo.toml
@@ -9,3 +9,5 @@ license = "MIT"
 [dependencies]
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+strum = "0.20.0"
+strum_macros = "0.20.0"
